Fix issue when `iptables_rules` table is not created for romana agent.
Romana agent uses locally managed database that used to be
created from a schema provided in pkg/util/firewall.

New storage won't respect it and will create database
full of useless in romana agent context tables.
And the only way to fix it is to add missing tables
to the list of useless in every other context tables.