Cleanup:
1. Moved some functionality into the client library, which now is used
by romanad's HTTP request handlers and also can be used by agent and
listener for the same functionality (e.g., adding/deleting a policy).

2. As we don't have Network IDs, and ExternalID vs Name vs ID were
endless cause of confusion (and internal ID was numeric just because
auto-incremented numeric ID works well in RDBMS), just left a single
ID field. Caller is responsible for making sure it's unique. Moreover,
since the tenant-network assignment is now part of topology
configuration, it is much easier to use a single ID provided by the
environment. We can add generating ID if missing, but that is not
needed for K8s case or indeed for OS case.

3. Moved various entities (tenant,segment,policy) from common to common/api,
so main common may be deprecated

4. Removed RestClient because nobody needs REST interface other than
CLI, and it can use go-resty as discussed. There's no rest-related
code to reuse from it anyway.